requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiofiles==24.1.0
    # via gradio
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anthropic==0.77.0
    # via -r requirements.in
anyio==4.12.1
    # via
    #   anthropic
    #   gradio
    #   httpx
    #   openai
    #   starlette
asttokens==3.0.1
    # via stack-data
brotli==1.2.0
    # via gradio
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   typer
    #   typer-slim
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   ipython
    #   tqdm
decorator==5.2.1
    # via ipython
distro==1.9.0
    # via
    #   anthropic
    #   openai
docstring-parser==0.17.0
    # via anthropic
executing==2.2.1
    # via stack-data
fastapi==0.128.0
    # via gradio
ffmpy==1.0.0
    # via gradio
filelock==3.20.3
    # via huggingface-hub
fsspec==2026.1.0
    # via
    #   gradio-client
    #   huggingface-hub
gradio==6.5.1
    # via -r requirements.in
gradio-client==2.0.3
    # via gradio
groovy==0.1.2
    # via gradio
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   safehttpx
huggingface-hub==1.3.5
    # via
    #   gradio
    #   gradio-client
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
ipython==9.9.0
    # via -r requirements.in
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via gradio
jiter==0.12.0
    # via
    #   anthropic
    #   openai
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   gradio
    #   jinja2
matplotlib-inline==0.2.1
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
numpy==2.4.1
    # via
    #   gradio
    #   pandas
openai==2.16.0
    # via -r requirements.in
orjson==3.11.6
    # via gradio
packaging==26.0
    # via
    #   gradio
    #   gradio-client
    #   huggingface-hub
pandas==3.0.0
    # via gradio
parso==0.8.5
    # via jedi
pillow==12.1.0
    # via gradio
prompt-toolkit==3.0.52
    # via ipython
pure-eval==0.2.3
    # via stack-data
pydantic==2.12.5
    # via
    #   anthropic
    #   fastapi
    #   gradio
    #   openai
pydantic-core==2.41.5
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   rich
pypdf==6.6.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.in
python-multipart==0.0.22
    # via gradio
pytz==2025.2
    # via gradio
pyyaml==6.0.3
    # via
    #   gradio
    #   huggingface-hub
requests==2.32.5
    # via -r requirements.in
rich==14.3.1
    # via
    #   -r requirements.in
    #   typer
safehttpx==0.1.7
    # via gradio
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   openai
stack-data==0.6.3
    # via ipython
starlette==0.50.0
    # via
    #   fastapi
    #   gradio
tomlkit==0.13.3
    # via gradio
tqdm==4.67.2
    # via
    #   huggingface-hub
    #   openai
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typer==0.21.1
    # via gradio
typer-slim==0.21.1
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   anthropic
    #   anyio
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typer
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
uvicorn==0.40.0
    # via gradio
wcwidth==0.5.3
    # via prompt-toolkit
